move ExpandableTagsQuery out of TagsQuery directory
