+++ src/exampleTypes/ExpandableTagsQuery/ExpandArrow.js
